requirements.txt
streamlit>=1.28.0
openai>=1.0.0
python-dotenv>=1.0.0
tqdm>=4.64.0
numpy>=1.21.0
